Changed appearance of AddForm component. Added new pics
