+++ src/components/AddForm/styles.module.css
@@ -1,0 +1,62 @@
+.container {
+  display: flex;
+  padding: 0 16px;
+  width: 100%;
+}
+
+.block {
+  display: flex;
+}
+
+.footer {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: space-between;
+}
+
+.userpicContainer {
+  padding: 12px 12px 0 0;
+}
+
+.buttonContainer {
+  margin-top: 8px;
+  border-radius: 20px;
+}
+
+.button {
+  height: 36px;
+  min-width: 36px;
+  text-align: center;
+  border-radius: 20px;
+  background-color: rgb(29, 155, 240);
+  border-color: rgba(0, 0, 0, 0);
+}
+
+.label {
+  margin: 15px;
+  color: white;
+}
+
+.icon {
+  height: 20px;
+  width: 20px;
+  padding-right: 5px;
+}
+
+.bottom-panel {
+  margin-top: 20px;
+  margin-bottom: 20px;
+}
+
+.bottom-panel .new-post-label {
+  width: auto;
+  flex-grow: 1;
+  margin-right: 3px;
+}
+
+.userpic {
+  height: 40px;
+  width: 40px;
+  border-radius: 200px;
+}
